[Markdown editor] Split plugin files
