+++ src/components/markdown_editor/plugins/markdown_default_plugins/parsing_plugins.ts
@@ -1,0 +1,41 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+// Importing seemingly unused types from `unified` because the definitions
+// are exported for two versions of TypeScript (3.4, 4.0) and implicit
+// imports during eui.d.ts generation default to the incorrect version (3.4).
+// Explicit imports here resolve the version mismatch.
+import {
+  PluggableList,
+  // @ts-ignore See above comment
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  Attacher,
+  // @ts-ignore See above comment
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  Pluggable,
+  // @ts-ignore See above comment
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  Settings,
+} from 'unified';
+import markdown from 'remark-parse';
+import emoji from 'remark-emoji';
+import highlight from '../remark/remark_prismjs';
+import * as MarkdownTooltip from '../markdown_tooltip';
+import * as MarkdownCheckbox from '../markdown_checkbox';
+import { markdownLinkValidator } from '../markdown_link_validator';
+
+export const getDefaultEuiMarkdownParsingPlugins = (): PluggableList => [
+  [markdown, {}],
+  [highlight, {}],
+  [emoji, { emoticon: true }],
+  [MarkdownTooltip.parser, {}],
+  [MarkdownCheckbox.parser, {}],
+  [markdownLinkValidator, {}],
+];
+
+export const defaultParsingPlugins = getDefaultEuiMarkdownParsingPlugins();
