+++ src/components/markdown_editor/plugins/markdown_tooltip/index.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+export { TooltipParser as parser } from './parser';
+export { tooltipPlugin as plugin } from './plugin';
+export { tooltipMarkdownRenderer as renderer } from './renderer';
+export * from './types';
